+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 - default empty file templates
 - fi, he, no, sl empty file templates
+- shardkey parameter
 
 ## Changed
 - demo server address changed